Refactor/HBAI-332: Refactor tool use (#71)
* HBAI-332 Use prompt template when inferencing.
Remove ragTemplate and ragMode.
Fix saving bot setting.
Fix loading bot setting.
Finalize agent class.
Finalize llama.cpp class.
Fix tool funcs.
Fix stopwords -> string only.

* HBAI-332 Upp ver

* HBAI-332 Add padding to splash page banner

* HBAI-332 Fixed error case when queue size exceeded.
Set "stop" default value to empty string.
Implement funcs for choosing tools.

* HBAI-332 Set startup window size to fraction of fullscreen

* HBAI-332 Fix tool example

* HBAI-332 Add mixtral model.
Fix descr text.

* HBAI-332 Remove uneeded code

* HBAI-332 Fix chat format for llama3

* HBAI-332 Fix tool choosing.
Set active role to default Agent.

* HBAI-332 Adjust window size

* HBAI-332 Add comments.
Add prompt formats.
Add model configs.

* HBAI-332 return a value

* HBAI-332 Dont override tool use command args

* HBAI-332 flush comments

* HBAI-332 Set view

* HBAI-332 Add gemma3 format

* HBAI-332 Fix second func call hanging

* HBAI-332 Update llama.cpp

* HBAI-332 Add return type to description.
Add to system prompt.

* HBAI-332 Remove comment.
Update prompt formats.

+++ backends/core/common.py
@@ -8,7 +8,7 @@ import httpx
 import subprocess
 from typing import Any, List, Optional, Tuple
 from core import classes
-from core.classes import InstalledTextModelMetadata, InstalledTextModel
+from core.classes import BotSettings, InstalledTextModelMetadata, InstalledTextModel
 from inference.classes import DEFAULT_MIN_CONTEXT_WINDOW, DEFAULT_CHAT_MODE, CHAT_MODES
 from huggingface_hub import (
     scan_cache_dir,
@@ -37,7 +37,10 @@ def dep_path(relative_path):
     return os.path.join(base_path, relative_path)
 
 
+CONTEXT_INPUT = "{context_str}"  # used by tools and RAG
+QUERY_INPUT = "{query_str}"
 MODEL_METADATAS_FILENAME = "installed_models.json"
+BACKENDS_FOLDER = "backends"
 APP_SETTINGS_FOLDER = "settings"
 APP_SETTINGS_PATH = app_path(APP_SETTINGS_FOLDER)
 TOOL_FOLDER = "tools"
@@ -478,7 +481,7 @@ def store_tool_definition(
                     os.remove(file_path)
 
 
-def save_bot_settings_file(folderpath: str, filepath: str, data: Any):
+def save_bot_settings_file(folderpath: str, filepath: str, data: BotSettings):
     # Create folder/file
     if not os.path.exists(folderpath):
         os.makedirs(folderpath)
@@ -494,11 +497,14 @@ def save_bot_settings_file(folderpath: str, filepath: str, data: Any):
         existing_data = []
 
     # Update the existing data
-    existing_data.append(data)
+    existing_data.append(data.model_dump())
 
     # Save the updated data to the file, this will overwrite all values
-    with open(filepath, "w") as file:
-        json.dump(existing_data, file, indent=2)
+    if isinstance(existing_data, list):
+        with open(filepath, "w") as file:
+            json.dump(existing_data, file, indent=2)
+    else:
+        print(f"{PRNT_API} Warning! Save data is malformed.", flush=True)
 
     return existing_data
